Backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.14
fastapi==0.109.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
numpy==2.4.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pypiwin32==223
python-dotenv==1.0.0
python-multipart==0.0.6
pyttsx3==2.99
pywin32==311
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
starlette==0.35.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.0
watchfiles==1.1.1
websockets==15.0.1
